Update deterministic masternode layer from dash/develop

+++ src/llmq/quorums_dkgsession.cpp
@@ -152,7 +152,7 @@ void CDKGSession::SendContributions(CDKGPendingMessages& pendingMessages)
     }
 
     CDKGContribution qc;
-    qc.llmqType = (uint8_t)params.type;
+    qc.llmqType = params.type;
     qc.quorumHash = pindexQuorum->GetBlockHash();
     qc.proTxHash = myProTxHash;
     qc.vvec = vvecContribution;
@@ -418,7 +418,7 @@ void CDKGSession::SendComplaint(CDKGPendingMessages& pendingMessages)
     assert(AreWeMember());
 
     CDKGComplaint qc(params);
-    qc.llmqType = (uint8_t)params.type;
+    qc.llmqType = params.type;
     qc.quorumHash = pindexQuorum->GetBlockHash();
     qc.proTxHash = myProTxHash;
 
@@ -601,7 +601,7 @@ void CDKGSession::SendJustification(CDKGPendingMessages& pendingMessages, const 
     LogPrint(BCLog::QUORUM, "sending justification for %d members", forMembers.size());
 
     CDKGJustification qj;
-    qj.llmqType = (uint8_t)params.type;
+    qj.llmqType = params.type;
     qj.quorumHash = pindexQuorum->GetBlockHash();
     qj.proTxHash = myProTxHash;
     qj.contributions.reserve(forMembers.size());
@@ -845,7 +845,7 @@ void CDKGSession::SendCommitment(CDKGPendingMessages& pendingMessages)
     LogPrint(BCLog::QUORUM, "sending commitment");
 
     CDKGPrematureCommitment qc(params);
-    qc.llmqType = (uint8_t)params.type;
+    qc.llmqType = params.type;
     qc.quorumHash = pindexQuorum->GetBlockHash();
     qc.proTxHash = myProTxHash;
